third_party/WebKit/LayoutTests/http/tests/inspector/network/request-name-path.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script>
function test()
{
    /**
     * @param {string} url
     * @param {string=} targetUrl
     */
    function createNetworkRequestForURLAndDumpNameAndPath(url, targetUrl)
    {
        var mainTarget = SDK.targetManager.mainTarget();
        var currentTargetURL = mainTarget.inspectedURL();
        if (targetUrl)
            mainTarget.setInspectedURL(targetUrl);
        InspectorTest.addResult("Dumping request name and path for url: " + url);
        var request = new SDK.NetworkRequest(mainTarget, 0, url);
        InspectorTest.addResult("    name = " + request.name());
        InspectorTest.addResult("    path = " + request.path());
        InspectorTest.addResult("    targetUrl = " + (targetUrl ? targetUrl : currentTargetURL));
        if (targetUrl !== currentTargetURL)
            mainTarget.setInspectedURL(currentTargetURL);
    }

    // Save the target URL to ensure test works well with other tests.
    var mainTarget = SDK.targetManager.mainTarget();
    var originalTargetURL = mainTarget.inspectedURL();
    mainTarget.setInspectedURL("http://127.0.0.1/aFolder/aTest.html");

    createNetworkRequestForURLAndDumpNameAndPath("http://www.example.com/foo/bar/baz?key=value");
    createNetworkRequestForURLAndDumpNameAndPath("http://www.example.com/foo/bar/?key=value");
    createNetworkRequestForURLAndDumpNameAndPath("http://www.example.com/foo/bar/baz");
    createNetworkRequestForURLAndDumpNameAndPath("http://www.example.com/foo/bar/");

    createNetworkRequestForURLAndDumpNameAndPath("http://www.example.com/baz?key=value");
    createNetworkRequestForURLAndDumpNameAndPath("http://www.example.com/?key=value");
    createNetworkRequestForURLAndDumpNameAndPath("http://www.example.com/baz");
    createNetworkRequestForURLAndDumpNameAndPath("http://www.example.com/");

    createNetworkRequestForURLAndDumpNameAndPath("http://127.0.0.1/foo/bar/baz?key=value");
    createNetworkRequestForURLAndDumpNameAndPath("http://127.0.0.1/foo/bar/?key=value");
    createNetworkRequestForURLAndDumpNameAndPath("http://127.0.0.1/foo/bar/baz");
    createNetworkRequestForURLAndDumpNameAndPath("http://127.0.0.1/foo/bar/");

    createNetworkRequestForURLAndDumpNameAndPath("http://127.0.0.1/baz?key=value");
    createNetworkRequestForURLAndDumpNameAndPath("http://127.0.0.1/?key=value");
    createNetworkRequestForURLAndDumpNameAndPath("http://127.0.0.1/baz");
    createNetworkRequestForURLAndDumpNameAndPath("http://127.0.0.1/");
    createNetworkRequestForURLAndDumpNameAndPath("http://127.0.0.1/", "data:text/html,<html><body><img src=\"data:,foobar\" /></body></html>");
    createNetworkRequestForURLAndDumpNameAndPath("data:text/html,<html><body><img src=\"data:,foobar\" /></body></html>", "data:text/html,<html><body><img src=\"data:,foobar\" /></body></html>");
    createNetworkRequestForURLAndDumpNameAndPath("data:text/html,<html><body><img src=\"data:,foobar\" /></body></html>", "http://127.0.0.1/");

    // Restore the target URL to ensure test works well with other tests.
    mainTarget.setInspectedURL(originalTargetURL);

    InspectorTest.completeTest();
}

</script>
</head>
<body onload="runTest()">
<p>Tests name() and path() methods of NetworkRequest.</p>
<a href="https://bugs.webkit.org/show_bug.cgi?id=101064">Bug 101064</a>
</body>
</html>
